BaseGeometry
